requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.13
alembic==1.13.2
appdirs==1.4.3
asn1crypto==0.24.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
bcrypt==3.1.3
blinker==1.6.2
build==1.2.1
certifi==2023.7.22
cffi==1.14.5
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.3
configparser==3.5.0
coverage==4.4.2
cryptography==42.0.4
dnspython==2.6.1
docutils==0.20.1
email_validator==2.2.0
enum34==1.1.6
flake8==3.3.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Mandrill==0.3
Flask-Migrate==3.0.1
Flask-Script==2.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
funcsigs==1.0.2
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
imagesize==1.4.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
libsql-client==0.3.1
Mako==1.2.2
MarkupSafe==2.1.1
mccabe==0.6.1
mock==2.0.0
multidict==6.0.5
nose2==0.6.5
packaging==24.1
pbr==3.1.1
pep8==1.7.0
pip-tools==7.4.1
pkg_resources==0.0.0
psycopg2-binary==2.9.1
pycodestyle==2.3.1
pycparser==2.17
pyflakes==1.5.0
Pygments==2.18.0
pyOpenSSL==19.0.0
pyparsing==2.2.0
pyproject_hooks==1.1.0
python-dotenv==0.17.1
python-editor==1.0.3
pytz==2024.1
requests==2.31.0
six==1.10.0
snowballstemmer==2.2.0
Sphinx==7.1.2
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.31
sqlalchemy-libsql==0.1.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.18
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2
